refactor: split app input test clusters
