+++ src/app/input/tests.rs
@@ -23,17 +23,6 @@ use crate::{
 use crossterm::event::{KeyModifiers, MouseEvent};
 use ratatui::layout::Rect;
 
-fn state_with_workspaces(names: &[&str]) -> AppState {
-    let mut state = AppState::test_new();
-    state.workspaces = names.iter().map(|name| Workspace::test_new(name)).collect();
-    if !state.workspaces.is_empty() {
-        state.active = Some(0);
-        state.selected = 0;
-        state.mode = Mode::Navigate;
-    }
-    state
-}
-
 fn app_for_mouse_test() -> App {
     let (_api_tx, api_rx) = tokio::sync::mpsc::unbounded_channel();
     let mut app = App::new(
@@ -79,580 +68,11 @@ fn capture_snapshot(state: &AppState) -> crate::persist::SessionSnapshot {
     )
 }
 
-fn unique_temp_path(name: &str) -> std::path::PathBuf {
-    let nanos = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map(|d| d.as_nanos())
-        .unwrap_or(0);
-    std::env::temp_dir().join(format!("herdr-{name}-{}-{nanos}", std::process::id()))
-}
-
-fn wait_for_file(path: &Path) -> String {
-    let deadline = std::time::Instant::now() + Duration::from_secs(2);
-    while std::time::Instant::now() < deadline {
-        if let Ok(content) = fs::read_to_string(path) {
-            return content;
-        }
-        std::thread::sleep(Duration::from_millis(20));
-    }
-    panic!("timed out waiting for {}", path.display());
-}
-
 fn root_layout_ratio(snapshot: &crate::persist::SessionSnapshot) -> Option<f32> {
     match &snapshot.workspaces.first()?.tabs.first()?.layout {
         crate::persist::LayoutSnapshot::Split { ratio, .. } => Some(*ratio),
         crate::persist::LayoutSnapshot::Pane(_) => None,
     }
-}
-
-#[test]
-fn custom_rename_key_enters_rename_mode() {
-    let mut state = state_with_workspaces(&["test"]);
-    state.keybinds.rename_workspace = (KeyCode::Char('g'), KeyModifiers::empty());
-    state.keybinds.rename_workspace_label = "g".into();
-
-    handle_navigate_key(
-        &mut state,
-        KeyEvent::new(KeyCode::Char('g'), KeyModifiers::empty()),
-    );
-
-    assert_eq!(state.mode, Mode::RenameWorkspace);
-    assert_eq!(state.name_input, "test");
-}
-
-#[test]
-fn custom_new_workspace_key_requests_and_exits_navigate() {
-    let mut state = state_with_workspaces(&["test"]);
-    state.keybinds.new_workspace = (KeyCode::Char('g'), KeyModifiers::empty());
-    state.keybinds.new_workspace_label = "g".into();
-
-    handle_navigate_key(
-        &mut state,
-        KeyEvent::new(KeyCode::Char('g'), KeyModifiers::empty()),
-    );
-
-    assert!(state.request_new_workspace);
-    assert_eq!(state.mode, Mode::Terminal);
-}
-
-#[test]
-fn custom_sidebar_toggle_key_toggles_and_exits_navigate() {
-    let mut state = state_with_workspaces(&["test"]);
-    state.keybinds.toggle_sidebar = (KeyCode::Char('g'), KeyModifiers::empty());
-    state.keybinds.toggle_sidebar_label = "g".into();
-    assert!(!state.sidebar_collapsed);
-
-    handle_navigate_key(
-        &mut state,
-        KeyEvent::new(KeyCode::Char('g'), KeyModifiers::empty()),
-    );
-
-    assert!(state.sidebar_collapsed);
-    assert_eq!(state.mode, Mode::Terminal);
-}
-
-#[test]
-fn custom_resize_key_enters_resize_mode() {
-    let mut state = state_with_workspaces(&["test"]);
-    state.keybinds.resize_mode = (KeyCode::Char('g'), KeyModifiers::empty());
-    state.keybinds.resize_mode_label = "g".into();
-
-    handle_navigate_key(
-        &mut state,
-        KeyEvent::new(KeyCode::Char('g'), KeyModifiers::empty()),
-    );
-
-    assert_eq!(state.mode, Mode::Resize);
-}
-
-#[test]
-fn movement_action_stays_in_navigate_mode() {
-    let mut state = state_with_workspaces(&["a", "b"]);
-    state.selected = 0;
-
-    handle_navigate_key(
-        &mut state,
-        KeyEvent::new(KeyCode::Down, KeyModifiers::empty()),
-    );
-
-    assert_eq!(state.selected, 1);
-    assert_eq!(state.mode, Mode::Navigate);
-}
-
-#[test]
-fn terminal_direct_focus_pane_shortcut_maps_to_navigation_action() {
-    let mut state = state_with_workspaces(&["test"]);
-    state.keybinds.focus_pane_left = Some((KeyCode::Left, KeyModifiers::ALT));
-    state.keybinds.focus_pane_left_label = Some("alt+left".into());
-
-    let action =
-        terminal_direct_navigation_action(&state, &KeyEvent::new(KeyCode::Left, KeyModifiers::ALT));
-
-    assert_eq!(action, Some(NavigateAction::FocusPaneLeft));
-}
-
-#[tokio::test]
-async fn terminal_direct_focus_pane_shortcut_switches_focus_without_leaving_terminal_mode() {
-    let (_api_tx, api_rx) = tokio::sync::mpsc::unbounded_channel();
-    let mut app = App::new(
-        &Config::default(),
-        true,
-        None,
-        None,
-        api_rx,
-        crate::api::EventHub::default(),
-    );
-    app.state.workspaces = vec![Workspace::test_new("test")];
-    app.state.active = Some(0);
-    app.state.selected = 0;
-    app.state.mode = Mode::Terminal;
-    app.state.workspaces[0].test_split(Direction::Horizontal);
-    app.state.view.pane_infos = app.state.workspaces[0]
-        .active_tab()
-        .unwrap()
-        .layout
-        .panes(Rect::new(0, 0, 80, 24));
-    let focused_before = app.state.workspaces[0].layout.focused();
-    app.state.keybinds.focus_pane_left = Some((KeyCode::Char('h'), KeyModifiers::ALT));
-    app.state.keybinds.focus_pane_left_label = Some("alt+h".into());
-
-    app.handle_terminal_key(TerminalKey::new(KeyCode::Char('h'), KeyModifiers::ALT))
-        .await;
-
-    assert_ne!(app.state.workspaces[0].layout.focused(), focused_before);
-    assert_eq!(app.state.mode, Mode::Terminal);
-}
-
-#[tokio::test]
-async fn custom_command_runs_from_prefix_key_in_navigate_mode() {
-    let (_api_tx, api_rx) = tokio::sync::mpsc::unbounded_channel();
-    let mut app = App::new(
-        &Config::default(),
-        true,
-        None,
-        None,
-        api_rx,
-        crate::api::EventHub::default(),
-    );
-    app.state.workspaces = vec![Workspace::test_new("test")];
-    app.state.active = Some(0);
-    app.state.selected = 0;
-    app.state.mode = Mode::Terminal;
-
-    let output_path = unique_temp_path("custom-command-keybind");
-    let command = format!(
-        "printf '%s\\n%s\\n%s\\n' \"$HERDR_ACTIVE_WORKSPACE_ID\" \"$HERDR_ACTIVE_TAB_ID\" \"$HERDR_ACTIVE_PANE_ID\" > '{}'",
-        output_path.display()
-    );
-    app.state.keybinds.custom_commands = vec![crate::config::CustomCommandKeybind {
-        key: (KeyCode::Char('g'), KeyModifiers::empty()),
-        label: "g".into(),
-        command,
-        action: crate::config::CustomCommandAction::Shell,
-    }];
-
-    app.handle_key(TerminalKey::new(
-        app.state.prefix_code,
-        app.state.prefix_mods,
-    ))
-    .await;
-    assert_eq!(app.state.mode, Mode::Navigate);
-
-    app.handle_key(TerminalKey::new(KeyCode::Char('g'), KeyModifiers::empty()))
-        .await;
-
-    let content = wait_for_file(&output_path);
-    let lines: Vec<&str> = content.lines().collect();
-    assert_eq!(lines.len(), 3);
-    assert_eq!(lines[0], app.state.workspaces[0].id);
-    assert_eq!(lines[1], format!("{}:1", app.state.workspaces[0].id));
-    assert_eq!(lines[2], format!("{}-1", app.state.workspaces[0].id));
-    assert_eq!(app.state.mode, Mode::Terminal);
-
-    let _ = fs::remove_file(output_path);
-}
-
-#[tokio::test]
-async fn pane_overlay_command_opens_and_closes_after_exit() {
-    let (_api_tx, api_rx) = tokio::sync::mpsc::unbounded_channel();
-    let mut app = App::new(
-        &Config::default(),
-        true,
-        None,
-        None,
-        api_rx,
-        crate::api::EventHub::default(),
-    );
-    let workspace = Workspace::new(
-        std::env::current_dir().unwrap_or_else(|_| "/".into()),
-        24,
-        80,
-        app.state.pane_scrollback_limit_bytes,
-        app.state.host_terminal_theme,
-        app.event_tx.clone(),
-        app.render_notify.clone(),
-        app.render_dirty.clone(),
-    )
-    .expect("workspace should spawn");
-    app.state.workspaces = vec![workspace];
-    app.state.active = Some(0);
-    app.state.selected = 0;
-    app.state.mode = Mode::Terminal;
-
-    let output_path = unique_temp_path("custom-pane-command");
-    let command = format!("printf done > '{}'", output_path.display());
-    app.state.keybinds.custom_commands = vec![crate::config::CustomCommandKeybind {
-        key: (KeyCode::Char('g'), KeyModifiers::empty()),
-        label: "g".into(),
-        command,
-        action: crate::config::CustomCommandAction::Pane,
-    }];
-
-    app.handle_key(TerminalKey::new(
-        app.state.prefix_code,
-        app.state.prefix_mods,
-    ))
-    .await;
-    app.handle_key(TerminalKey::new(KeyCode::Char('g'), KeyModifiers::empty()))
-        .await;
-
-    assert_eq!(app.state.workspaces[0].tabs[0].layout.pane_count(), 2);
-    assert!(app.state.workspaces[0].tabs[0].zoomed);
-
-    let _ = wait_for_file(&output_path);
-    let deadline = std::time::Instant::now() + Duration::from_secs(2);
-    while std::time::Instant::now() < deadline {
-        if app.drain_internal_events() && app.state.workspaces[0].tabs[0].layout.pane_count() == 1 {
-            break;
-        }
-        std::thread::sleep(Duration::from_millis(20));
-    }
-
-    assert_eq!(app.state.workspaces[0].tabs[0].layout.pane_count(), 1);
-    assert!(!app.state.workspaces[0].tabs[0].zoomed);
-    assert_eq!(app.state.mode, Mode::Terminal);
-    let _ = fs::remove_file(output_path);
-}
-
-#[test]
-fn fullscreen_action_exits_navigate_mode() {
-    let mut state = state_with_workspaces(&["test"]);
-    state.workspaces[0].test_split(Direction::Horizontal);
-    state.keybinds.fullscreen = (KeyCode::Char('g'), KeyModifiers::empty());
-    state.keybinds.fullscreen_label = "g".into();
-
-    handle_navigate_key(
-        &mut state,
-        KeyEvent::new(KeyCode::Char('g'), KeyModifiers::empty()),
-    );
-
-    assert!(state.workspaces[0].zoomed);
-    assert_eq!(state.mode, Mode::Terminal);
-}
-
-#[test]
-fn custom_resize_key_exits_resize_mode() {
-    let mut state = state_with_workspaces(&["test"]);
-    state.mode = Mode::Resize;
-    state.keybinds.resize_mode = (KeyCode::Char('g'), KeyModifiers::empty());
-    state.keybinds.resize_mode_label = "g".into();
-
-    handle_resize_key(
-        &mut state,
-        KeyEvent::new(KeyCode::Char('g'), KeyModifiers::empty()),
-    );
-
-    assert_eq!(state.mode, Mode::Terminal);
-}
-
-#[test]
-fn settings_cancel_restores_previewed_theme_from_other_sections() {
-    let mut state = state_with_workspaces(&["test"]);
-    let original_palette = state.palette.clone();
-    let original_theme = state.theme_name.clone();
-
-    open_settings(&mut state);
-    update_settings_state(
-        &mut state,
-        KeyEvent::new(KeyCode::Down, KeyModifiers::empty()),
-    );
-    assert_ne!(state.theme_name, original_theme);
-
-    update_settings_state(
-        &mut state,
-        KeyEvent::new(KeyCode::Tab, KeyModifiers::empty()),
-    );
-    assert_eq!(
-        state.settings.section,
-        crate::app::state::SettingsSection::Sound
-    );
-
-    update_settings_state(
-        &mut state,
-        KeyEvent::new(KeyCode::Esc, KeyModifiers::empty()),
-    );
-
-    assert_eq!(state.mode, Mode::Terminal);
-    assert_eq!(state.theme_name, original_theme);
-    assert_eq!(state.palette.accent, original_palette.accent);
-    assert_eq!(state.palette.panel_bg, original_palette.panel_bg);
-}
-
-#[test]
-fn settings_sound_toggle_returns_save_action() {
-    let mut state = state_with_workspaces(&["test"]);
-    open_settings(&mut state);
-    state.settings.section = crate::app::state::SettingsSection::Sound;
-    state.settings.list.selected = 0;
-
-    let action = update_settings_state(
-        &mut state,
-        KeyEvent::new(KeyCode::Enter, KeyModifiers::empty()),
-    );
-
-    assert_eq!(action, Some(SettingsAction::SaveSound(true)));
-    assert!(state.sound.enabled);
-    assert_eq!(state.mode, Mode::Settings);
-}
-
-#[test]
-fn question_mark_opens_keybind_help_from_navigate() {
-    let mut state = state_with_workspaces(&["test"]);
-
-    handle_navigate_key(
-        &mut state,
-        KeyEvent::new(KeyCode::Char('?'), KeyModifiers::SHIFT),
-    );
-
-    assert_eq!(state.mode, Mode::KeybindHelp);
-}
-
-#[test]
-fn rename_modal_keyboard_and_mouse_share_actions() {
-    let mut state = state_with_workspaces(&["test"]);
-    state.mode = Mode::RenameWorkspace;
-    state.name_input = "hello".into();
-
-    handle_rename_key(
-        &mut state,
-        KeyEvent::new(KeyCode::Char('c'), KeyModifiers::CONTROL),
-    );
-    assert!(state.name_input.is_empty());
-
-    state.name_input = "renamed".into();
-    handle_rename_key(
-        &mut state,
-        KeyEvent::new(KeyCode::Enter, KeyModifiers::empty()),
-    );
-    assert_eq!(state.mode, Mode::Terminal);
-    assert_eq!(state.workspaces[0].display_name(), "renamed");
-    let snapshot = capture_snapshot(&state);
-    assert_eq!(
-        snapshot.workspaces[0].custom_name.as_deref(),
-        Some("renamed")
-    );
-
-    state.view.sidebar_rect = Rect::new(0, 0, 26, 20);
-    state.view.terminal_area = Rect::new(26, 0, 80, 20);
-    state.mode = Mode::RenameWorkspace;
-    state.name_input = "mouse".into();
-    let inner = state.rename_modal_inner().unwrap();
-    let (save, _, _) = crate::ui::rename_button_rects(inner);
-    let action = modal_action_from_buttons(save.x, save.y, &[(save, ModalAction::Save)]);
-    assert_eq!(action, Some(ModalAction::Save));
-}
-
-#[test]
-fn tab_rename_updates_captured_snapshot() {
-    let mut state = state_with_workspaces(&["test"]);
-    state.mode = Mode::RenameTab;
-    state.name_input = "logs".into();
-
-    handle_rename_key(
-        &mut state,
-        KeyEvent::new(KeyCode::Enter, KeyModifiers::empty()),
-    );
-
-    let snapshot = capture_snapshot(&state);
-    assert_eq!(
-        snapshot.workspaces[0].tabs[0].custom_name.as_deref(),
-        Some("logs")
-    );
-}
-
-#[test]
-fn rename_cancel_returns_to_terminal_when_workspace_is_active() {
-    let mut state = state_with_workspaces(&["test"]);
-    state.mode = Mode::RenameTab;
-    state.name_input = "test".into();
-
-    handle_rename_key(
-        &mut state,
-        KeyEvent::new(KeyCode::Esc, KeyModifiers::empty()),
-    );
-
-    assert_eq!(state.mode, Mode::Terminal);
-    assert!(state.name_input.is_empty());
-}
-
-#[test]
-fn rename_modal_replaces_prefilled_text_on_first_type() {
-    let mut state = state_with_workspaces(&["test"]);
-    state.mode = Mode::RenameTab;
-    state.name_input = "2".into();
-    state.name_input_replace_on_type = true;
-
-    handle_rename_key(
-        &mut state,
-        KeyEvent::new(KeyCode::Char('n'), KeyModifiers::empty()),
-    );
-    assert_eq!(state.name_input, "n");
-    assert!(!state.name_input_replace_on_type);
-
-    handle_rename_key(
-        &mut state,
-        KeyEvent::new(KeyCode::Char('e'), KeyModifiers::empty()),
-    );
-    assert_eq!(state.name_input, "ne");
-}
-
-#[test]
-fn open_rename_active_tab_can_prefill_default_new_tab_name() {
-    let mut state = state_with_workspaces(&["test"]);
-    state.workspaces[0].test_add_tab(None);
-    state.workspaces[0].switch_tab(1);
-
-    open_rename_active_tab(&mut state, true);
-
-    assert_eq!(state.mode, Mode::RenameTab);
-    assert_eq!(state.name_input, "2");
-    assert!(state.name_input_replace_on_type);
-}
-
-#[test]
-fn new_tab_action_opens_dialog_without_creating_tab() {
-    let mut state = state_with_workspaces(&["test"]);
-
-    execute_navigate_action(&mut state, NavigateAction::NewTab);
-
-    assert_eq!(state.mode, Mode::RenameTab);
-    assert!(state.creating_new_tab);
-    assert_eq!(state.name_input, "2");
-    assert!(state.name_input_replace_on_type);
-    assert!(!state.request_new_tab);
-    assert_eq!(state.workspaces[0].tabs.len(), 1);
-}
-
-#[test]
-fn cancel_new_tab_dialog_leaves_workspace_unchanged() {
-    let mut state = state_with_workspaces(&["test"]);
-    open_new_tab_dialog(&mut state);
-
-    handle_rename_key(
-        &mut state,
-        KeyEvent::new(KeyCode::Esc, KeyModifiers::empty()),
-    );
-
-    assert_eq!(state.mode, Mode::Terminal);
-    assert!(!state.creating_new_tab);
-    assert!(!state.request_new_tab);
-    assert!(state.requested_new_tab_name.is_none());
-    assert_eq!(state.workspaces[0].tabs.len(), 1);
-}
-
-#[test]
-fn saving_new_tab_dialog_requests_creation_with_name() {
-    let mut state = state_with_workspaces(&["test"]);
-    open_new_tab_dialog(&mut state);
-    state.name_input = "logs".into();
-    state.name_input_replace_on_type = false;
-
-    handle_rename_key(
-        &mut state,
-        KeyEvent::new(KeyCode::Enter, KeyModifiers::empty()),
-    );
-
-    assert_eq!(state.mode, Mode::Terminal);
-    assert!(!state.creating_new_tab);
-    assert!(state.request_new_tab);
-    assert_eq!(state.requested_new_tab_name.as_deref(), Some("logs"));
-}
-
-#[test]
-fn saving_new_tab_dialog_with_default_name_keeps_tab_auto_named() {
-    let mut state = state_with_workspaces(&["test"]);
-    open_new_tab_dialog(&mut state);
-
-    handle_rename_key(
-        &mut state,
-        KeyEvent::new(KeyCode::Enter, KeyModifiers::empty()),
-    );
-
-    assert_eq!(state.mode, Mode::Terminal);
-    assert!(!state.creating_new_tab);
-    assert!(state.request_new_tab);
-    assert!(state.requested_new_tab_name.is_none());
-}
-
-#[test]
-fn closing_first_auto_tab_resets_remaining_auto_tab_and_next_prompt() {
-    let mut state = state_with_workspaces(&["test"]);
-    open_new_tab_dialog(&mut state);
-    handle_rename_key(
-        &mut state,
-        KeyEvent::new(KeyCode::Enter, KeyModifiers::empty()),
-    );
-
-    state.workspaces[0].test_add_tab(state.requested_new_tab_name.as_deref());
-    state.request_new_tab = false;
-    state.requested_new_tab_name = None;
-
-    state.workspaces[0].close_tab(0);
-    state.workspaces[0].switch_tab(0);
-
-    assert_eq!(state.workspaces[0].tabs[0].display_name(), "1");
-    assert!(state.workspaces[0].tabs[0].custom_name.is_none());
-
-    open_new_tab_dialog(&mut state);
-    assert_eq!(state.name_input, "2");
-}
-
-#[test]
-fn renaming_auto_tab_to_its_default_number_keeps_it_auto_named() {
-    let mut state = state_with_workspaces(&["test"]);
-    state.workspaces[0].test_add_tab(None);
-    state.workspaces[0].switch_tab(1);
-
-    open_rename_active_tab(&mut state, false);
-    handle_rename_key(
-        &mut state,
-        KeyEvent::new(KeyCode::Enter, KeyModifiers::empty()),
-    );
-
-    assert_eq!(state.mode, Mode::Terminal);
-    assert!(state.workspaces[0].tabs[1].custom_name.is_none());
-    assert_eq!(state.workspaces[0].tabs[1].display_name(), "2");
-}
-
-#[test]
-fn confirm_close_keyboard_actions_are_direct_not_focused() {
-    let mut state = state_with_workspaces(&["a", "b"]);
-    state.mode = Mode::ConfirmClose;
-    state.selected = 1;
-
-    handle_confirm_close_key(
-        &mut state,
-        KeyEvent::new(KeyCode::Esc, KeyModifiers::empty()),
-    );
-    assert_eq!(state.mode, Mode::Navigate);
-    assert_eq!(state.workspaces.len(), 2);
-
-    state.mode = Mode::ConfirmClose;
-    handle_confirm_close_key(
-        &mut state,
-        KeyEvent::new(KeyCode::Enter, KeyModifiers::empty()),
-    );
-    assert_eq!(state.workspaces.len(), 1);
 }
 
 #[test]
@@ -799,19 +219,6 @@ fn persistence_mode_menu_surfaces_detach_action() {
 
     assert!(app.state.detach_requested);
     assert!(!app.state.should_quit);
-}
-
-#[test]
-fn persistence_mode_navigate_q_detaches_instead_of_quitting_server() {
-    let mut state = AppState::test_new();
-    state.quit_detaches = true;
-
-    assert!(handle_navigate_reserved_key(
-        &mut state,
-        KeyEvent::new(KeyCode::Char('q'), KeyModifiers::empty())
-    ));
-    assert!(state.detach_requested);
-    assert!(!state.should_quit);
 }
 
 #[test]
@@ -1244,166 +651,6 @@ fn clicking_collapsed_sidebar_toggle_expands_sidebar() {
     assert!(!app.state.sidebar_collapsed);
 }
 
-#[tokio::test]
-async fn dragging_selection_above_pane_autoscrolls_and_extends_into_scrollback() {
-    let mut app = app_for_mouse_test();
-    let mut ws = Workspace::test_new("test");
-    let pane_id = ws.tabs[0].root_pane;
-    let pane_infos = ws.tabs[0].layout.panes(Rect::new(26, 2, 80, 18));
-    let info = pane_infos[0].clone();
-    ws.tabs[0].runtimes.insert(
-        pane_id,
-        crate::pane::PaneRuntime::test_with_scrollback_bytes(
-            info.inner_rect.width,
-            info.inner_rect.height,
-            16 * 1024,
-            &numbered_lines_bytes(64),
-        ),
-    );
-
-    app.state.workspaces = vec![ws];
-    app.state.active = Some(0);
-    app.state.selected = 0;
-    app.state.mode = Mode::Terminal;
-    app.state.view.pane_infos = pane_infos;
-
-    let start_metrics = app.state.workspaces[0]
-        .runtime(pane_id)
-        .and_then(crate::pane::PaneRuntime::scroll_metrics)
-        .expect("initial scroll metrics");
-    let start_row = info.inner_rect.y;
-    let start_col = info.inner_rect.x + 2;
-
-    app.handle_mouse(mouse(
-        MouseEventKind::Down(MouseButton::Left),
-        start_col,
-        start_row,
-    ));
-    app.handle_mouse(mouse(
-        MouseEventKind::Drag(MouseButton::Left),
-        start_col,
-        info.inner_rect.y.saturating_sub(1),
-    ));
-
-    let end_metrics = app.state.workspaces[0]
-        .runtime(pane_id)
-        .and_then(crate::pane::PaneRuntime::scroll_metrics)
-        .expect("scroll metrics after drag");
-    assert_eq!(
-        end_metrics.offset_from_bottom,
-        start_metrics.offset_from_bottom + 3
-    );
-
-    let selection = app.state.selection.as_ref().expect("selection after drag");
-    assert!(selection.is_visible());
-    assert_eq!(
-        selection.ordered_cells(),
-        (
-            (
-                (start_metrics.max_offset_from_bottom - end_metrics.offset_from_bottom) as u32,
-                2,
-            ),
-            (start_metrics.max_offset_from_bottom as u32, 2),
-        )
-    );
-}
-
-#[tokio::test]
-async fn releasing_dragged_selection_clears_highlight_after_copy() {
-    let mut app = app_for_mouse_test();
-    let mut ws = Workspace::test_new("test");
-    let pane_id = ws.tabs[0].root_pane;
-    let pane_infos = ws.tabs[0].layout.panes(Rect::new(26, 2, 80, 18));
-    let info = pane_infos[0].clone();
-    ws.tabs[0].runtimes.insert(
-        pane_id,
-        crate::pane::PaneRuntime::test_with_scrollback_bytes(
-            info.inner_rect.width,
-            info.inner_rect.height,
-            16 * 1024,
-            &numbered_lines_bytes(64),
-        ),
-    );
-
-    app.state.workspaces = vec![ws];
-    app.state.active = Some(0);
-    app.state.selected = 0;
-    app.state.mode = Mode::Terminal;
-    app.state.view.pane_infos = pane_infos;
-
-    let row = info.inner_rect.y;
-    let start_col = info.inner_rect.x + 1;
-    let end_col = info.inner_rect.x + 4;
-
-    app.handle_mouse(mouse(
-        MouseEventKind::Down(MouseButton::Left),
-        start_col,
-        row,
-    ));
-    app.handle_mouse(mouse(MouseEventKind::Drag(MouseButton::Left), end_col, row));
-    assert!(app.state.selection.is_some());
-
-    app.handle_mouse(mouse(MouseEventKind::Up(MouseButton::Left), end_col, row));
-
-    assert!(app.state.selection.is_none());
-}
-
-#[tokio::test]
-async fn wheel_scroll_keeps_in_progress_selection_and_extends_it() {
-    let mut app = app_for_mouse_test();
-    let mut ws = Workspace::test_new("test");
-    let pane_id = ws.tabs[0].root_pane;
-    let pane_infos = ws.tabs[0].layout.panes(Rect::new(26, 2, 80, 18));
-    let info = pane_infos[0].clone();
-    ws.tabs[0].runtimes.insert(
-        pane_id,
-        crate::pane::PaneRuntime::test_with_scrollback_bytes(
-            info.inner_rect.width,
-            info.inner_rect.height,
-            16 * 1024,
-            &numbered_lines_bytes(64),
-        ),
-    );
-
-    app.state.workspaces = vec![ws];
-    app.state.active = Some(0);
-    app.state.selected = 0;
-    app.state.mode = Mode::Terminal;
-    app.state.view.pane_infos = pane_infos;
-
-    let start_metrics = app.state.workspaces[0]
-        .runtime(pane_id)
-        .and_then(crate::pane::PaneRuntime::scroll_metrics)
-        .expect("initial scroll metrics");
-    let top_row = info.inner_rect.y;
-    let col = info.inner_rect.x + 2;
-
-    app.handle_mouse(mouse(MouseEventKind::Down(MouseButton::Left), col, top_row));
-    app.handle_mouse(mouse(MouseEventKind::ScrollUp, col, top_row));
-
-    let end_metrics = app.state.workspaces[0]
-        .runtime(pane_id)
-        .and_then(crate::pane::PaneRuntime::scroll_metrics)
-        .expect("scroll metrics after wheel");
-    assert_eq!(
-        end_metrics.offset_from_bottom,
-        start_metrics.offset_from_bottom + 3
-    );
-
-    let selection = app.state.selection.as_ref().expect("selection after wheel");
-    assert!(selection.is_visible());
-    assert_eq!(
-        selection.ordered_cells(),
-        (
-            (
-                (start_metrics.max_offset_from_bottom - end_metrics.offset_from_bottom) as u32,
-                2,
-            ),
-            (start_metrics.max_offset_from_bottom as u32, 2),
-        )
-    );
-}
-
 #[test]
 fn clicking_workspace_switches_on_mouse_up() {
     let mut app = app_for_mouse_test();
@@ -1776,142 +1023,4 @@ fn wheel_routing_falls_back_to_host_scrollback() {
     };
 
     assert_eq!(wheel_routing(input_state), WheelRouting::HostScroll);
-}
-
-#[tokio::test]
-async fn clicking_unfocused_pane_with_mouse_reporting_focuses_it_via_left_button() {
-    let mut app = app_for_mouse_test();
-    let mut ws = Workspace::test_new("test");
-    let first_pane = ws.tabs[0].root_pane;
-    let second_pane = ws.test_split(ratatui::layout::Direction::Vertical);
-
-    let terminal_area = Rect::new(26, 2, 80, 18);
-    let pane_infos = ws.tabs[0].layout.panes(terminal_area);
-    let first_info = pane_infos
-        .iter()
-        .find(|p| p.id == first_pane)
-        .expect("first pane info")
-        .clone();
-    let second_info = pane_infos
-        .iter()
-        .find(|p| p.id == second_pane)
-        .expect("second pane info")
-        .clone();
-
-    ws.tabs[0].runtimes.insert(
-        first_pane,
-        crate::pane::PaneRuntime::test_with_screen_bytes(
-            first_info.inner_rect.width.max(1),
-            first_info.inner_rect.height.max(1),
-            b"",
-        ),
-    );
-    ws.tabs[0].runtimes.insert(
-        second_pane,
-        crate::pane::PaneRuntime::test_with_screen_bytes(
-            second_info.inner_rect.width.max(1),
-            second_info.inner_rect.height.max(1),
-            b"\x1b[?1002h",
-        ),
-    );
-
-    ws.tabs[0].layout.focus_pane(first_pane);
-
-    app.state.workspaces = vec![ws];
-    app.state.active = Some(0);
-    app.state.selected = 0;
-    app.state.mode = Mode::Terminal;
-    app.state.view.pane_infos = pane_infos;
-
-    assert_eq!(
-        app.state.workspaces[0].tabs[0].layout.focused(),
-        first_pane,
-        "first pane should be focused before click"
-    );
-
-    app.handle_mouse(mouse(
-        MouseEventKind::Down(MouseButton::Left),
-        second_info.inner_rect.x + 2,
-        second_info.inner_rect.y + 2,
-    ));
-
-    assert_eq!(
-        app.state.workspaces[0].tabs[0].layout.focused(),
-        second_pane,
-        "left-clicking a pane with mouse reporting should move focus to it"
-    );
-    assert_eq!(app.state.mode, Mode::Terminal);
-}
-
-#[tokio::test]
-async fn clicking_unfocused_pane_with_mouse_reporting_focuses_it_via_right_button() {
-    let mut app = app_for_mouse_test();
-    let mut ws = Workspace::test_new("test");
-    let first_pane = ws.tabs[0].root_pane;
-    let second_pane = ws.test_split(ratatui::layout::Direction::Vertical);
-
-    let terminal_area = Rect::new(26, 2, 80, 18);
-    let pane_infos = ws.tabs[0].layout.panes(terminal_area);
-    let first_info = pane_infos
-        .iter()
-        .find(|p| p.id == first_pane)
-        .expect("first pane info")
-        .clone();
-    let second_info = pane_infos
-        .iter()
-        .find(|p| p.id == second_pane)
-        .expect("second pane info")
-        .clone();
-
-    ws.tabs[0].runtimes.insert(
-        first_pane,
-        crate::pane::PaneRuntime::test_with_screen_bytes(
-            first_info.inner_rect.width.max(1),
-            first_info.inner_rect.height.max(1),
-            b"",
-        ),
-    );
-    ws.tabs[0].runtimes.insert(
-        second_pane,
-        crate::pane::PaneRuntime::test_with_screen_bytes(
-            second_info.inner_rect.width.max(1),
-            second_info.inner_rect.height.max(1),
-            b"\x1b[?1002h",
-        ),
-    );
-
-    ws.tabs[0].layout.focus_pane(first_pane);
-
-    app.state.workspaces = vec![ws];
-    app.state.active = Some(0);
-    app.state.selected = 0;
-    app.state.mode = Mode::Terminal;
-    app.state.view.pane_infos = pane_infos;
-
-    assert_eq!(
-        app.state.workspaces[0].tabs[0].layout.focused(),
-        first_pane,
-        "first pane should be focused before click"
-    );
-
-    app.handle_mouse(mouse(
-        MouseEventKind::Down(MouseButton::Right),
-        second_info.inner_rect.x + 2,
-        second_info.inner_rect.y + 2,
-    ));
-
-    assert_eq!(
-        app.state.workspaces[0].tabs[0].layout.focused(),
-        second_pane,
-        "right-clicking a pane with mouse reporting should move focus to it"
-    );
-    assert_eq!(
-        app.state.mode,
-        Mode::ContextMenu,
-        "right-click should enter ContextMenu mode"
-    );
-    assert!(
-        app.state.context_menu.is_some(),
-        "right-click should populate context_menu state"
-    );
 }
